Display messages when saving or deleting entries.
The messages are baked into the DOM to make displaying them easier, and are hidden and shown as needed. This implementation is slightly race-y - multiple edits to the same field can cause the "saving" message to disappear early - but this is harmless as this mostly exists to tell users that it's okay to navigate away from the page without an explicit "save" button.

I've deliberately omitted any attempt to display a message for new (blank) entries, as the user doesn't actually care if we save them (and we don't display them anyways).

Fixes #12.