Allow reordering contacts.
This uses jQuery UI (which unfortunately does not work with jQuery Slim) to provide drag-and-drop ordering. The resulting list of IDs is passed to an API endpoint that handles ordering.

On the backend, the theory is that the newly-added seq value for each internetidentity allows an _optional_ ordering, with the current "creation order" ordering governing any values that don't have a seq. This includes new entries, as well as everything that exists when this is deployed.

The API endpoint does two things:
* Nulls out all seq values for the current user, to forget previous ordering assignments, and
* Sets seq to a serial value for each internetidentity.

Fixes #11.