Add the vendor chunk to the base template
I am not really happy with this solution; in dev, we get spurious errors
about css/app.css and js/chunk-vendors.js being missing, when they are
not needed, and thus not generated. I'd like dev and production setups
to be as close to each other as possible, ideally. But for now, this
should get things working at least.