Spike out alias/search API/UI
This does not include tests, and breaks a few existing tests due to
phrasing changes.

This also has a lot of duplication between EditIdentityForm and
EditAliasForm. That needs cleaning up.

This also really forces us to confront the "attach Vue to a bit of
pre-rendered DOM" issue. Lots of duplication between Django templates
and Vue components, which means they start to get more and more out of
sync. This needs to be addressed.